extract regenerate-common.ts
